+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.2
 
+- renamed gem from g5-dbupdate to g5_dbupdate
 - Added `gemfury_helpers` gem to `rake release` into Gemfury
 - Added `.ruby-version` (2.2.1)